chore: move config to internal package
