+++ config.go
@@ -2,30 +2,23 @@ package testcontainers
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"path/filepath"
-	"strconv"
 	"sync"
 
-	"github.com/magiconair/properties"
-	"github.com/testcontainers/testcontainers-go/internal/testcontainersdocker"
+	"github.com/testcontainers/testcontainers-go/internal/config"
 )
 
 var tcConfig TestcontainersConfig
 var tcConfigOnce *sync.Once = new(sync.Once)
 
 // TestcontainersConfig represents the configuration for Testcontainers
-// testcontainersConfig {
 type TestcontainersConfig struct {
-	Host           string `properties:"docker.host,default="`
-	TLSVerify      int    `properties:"docker.tls.verify,default=0"`
-	CertPath       string `properties:"docker.cert.path,default="`
-	RyukDisabled   bool   `properties:"ryuk.disabled,default=false"`
-	RyukPrivileged bool   `properties:"ryuk.container.privileged,default=false"`
+	Host           string `properties:"docker.host,default="`                    // Deprecated: use Config.Host instead
+	TLSVerify      int    `properties:"docker.tls.verify,default=0"`             // Deprecated: use Config.TLSVerify instead
+	CertPath       string `properties:"docker.cert.path,default="`               // Deprecated: use Config.CertPath instead
+	RyukDisabled   bool   `properties:"ryuk.disabled,default=false"`             // Deprecated: use Config.RyukDisabled instead
+	RyukPrivileged bool   `properties:"ryuk.container.privileged,default=false"` // Deprecated: use Config.RyukPrivileged instead
+	Config         config.Config
 }
-
-// }
 
 // ReadConfig reads from testcontainers properties file, storing the result in a singleton instance
 // of the TestcontainersConfig struct
@@ -38,73 +31,20 @@ func ReadConfig() TestcontainersConfig {
 // of the TestcontainersConfig struct
 func ReadConfigWithContext(ctx context.Context) TestcontainersConfig {
 	tcConfigOnce.Do(func() {
-		tcConfig = readConfig(ctx)
+		cfg := config.Read(ctx)
 
-		if tcConfig.RyukDisabled {
+		tcConfig.Config = cfg
+
+		if cfg.RyukDisabled {
 			ryukDisabledMessage := `
 **********************************************************************************************
 Ryuk has been disabled for the current execution. This can cause unexpected behavior in your environment.
 More on this: https://golang.testcontainers.org/features/garbage_collector/
 **********************************************************************************************`
 			Logger.Printf(ryukDisabledMessage)
-			Logger.Printf("\n%+v", tcConfig)
+			Logger.Printf("\n%+v", cfg)
 		}
 	})
 
 	return tcConfig
-}
-
-// readConfig reads from testcontainers properties file, if it exists
-// it is possible that certain values get overridden when set as environment variables
-func readConfig(ctx context.Context) TestcontainersConfig {
-	config := TestcontainersConfig{}
-
-	applyEnvironmentConfiguration := func(config TestcontainersConfig) TestcontainersConfig {
-		if dockerHostEnv := testcontainersdocker.ExtractDockerHost(ctx); dockerHostEnv != "" {
-			config.Host = dockerHostEnv
-		}
-		if config.Host == "" {
-			config.Host = testcontainersdocker.DockerSocketPathWithSchema
-		}
-
-		ryukDisabledEnv := os.Getenv("TESTCONTAINERS_RYUK_DISABLED")
-		if parseBool(ryukDisabledEnv) {
-			config.RyukDisabled = ryukDisabledEnv == "true"
-		}
-
-		ryukPrivilegedEnv := os.Getenv("TESTCONTAINERS_RYUK_CONTAINER_PRIVILEGED")
-		if parseBool(ryukPrivilegedEnv) {
-			config.RyukPrivileged = ryukPrivilegedEnv == "true"
-		}
-
-		return config
-	}
-
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return applyEnvironmentConfiguration(config)
-	}
-
-	tcProp := filepath.Join(home, ".testcontainers.properties")
-	// init from a file
-	properties, err := properties.LoadFile(tcProp, properties.UTF8)
-	if err != nil {
-		return applyEnvironmentConfiguration(config)
-	}
-
-	if err := properties.Decode(&config); err != nil {
-		fmt.Printf("invalid testcontainers properties file, returning an empty Testcontainers configuration: %v\n", err)
-		return applyEnvironmentConfiguration(config)
-	}
-
-	fmt.Printf("Testcontainers properties file has been found: %s\n", tcProp)
-
-	return applyEnvironmentConfiguration(config)
-}
-
-func parseBool(input string) bool {
-	if _, err := strconv.ParseBool(input); err == nil {
-		return true
-	}
-	return false
 }
