+++ reaper.go
@@ -73,8 +73,7 @@ func reuseOrCreateReaper(ctx context.Context, sessionID string, provider ReaperP
 // newReaper creates a Reaper with a sessionID to identify containers and a provider to use
 // Should only be used internally and instead use reuseOrCreateReaper to prefer reusing an existing Reaper instance
 func newReaper(ctx context.Context, sessionID string, provider ReaperProvider, opts ...ContainerOption) (*Reaper, error) {
-	// read host from config, which has been loaded once
-	dockerHost := ReadConfigWithContext(ctx).Host
+	dockerHost := testcontainersdocker.ExtractDockerHost(ctx)
 
 	reaper := &Reaper{
 		Provider:  provider,
@@ -83,7 +82,7 @@ func newReaper(ctx context.Context, sessionID string, provider ReaperProvider, o
 
 	listeningPort := nat.Port("8080/tcp")
 
-	tcConfig := provider.Config()
+	tcConfig := provider.Config().Config
 
 	reaperOpts := containerOptions{}
